Rename Toggle component file
